Move writeRotate impl
